# server/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuário autenticado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "access_token": {
                      "type": "string"
                    },
                    "user": {
                      "$ref": "#/components/schemas/UserEntity"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Credenciais inválidas"
          }
        },
        "summary": "Autenticar usuário",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário registrado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserEntity"
                }
              }
            }
          }
        },
        "summary": "Registrar novo usuário",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/profile": {
      "get": {
        "operationId": "AuthController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Perfil do usuário retornado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obter perfil do usuário autenticado",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/change-password": {
      "patch": {
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Senha alterada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Alterar senha do usuário autenticado",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Futebasssss API",
    "description": "API para o sistema Futebasssss",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "UserEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "ID do usuário",
            "example": 1
          },
          "email": {
            "type": "string",
            "description": "Email do usuário",
            "example": "usuario@exemplo.com"
          },
          "name": {
            "type": "string",
            "description": "Nome do usuário",
            "example": "Nome do Usuário"
          },
          "type": {
            "type": "string",
            "description": "Tipo de usuário",
            "example": "PLAYER",
            "enum": [
              "PLAYER",
              "ADMIN"
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Data de criação do usuário",
            "example": "2023-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Data de atualização do usuário",
            "example": "2023-01-01T00:00:00.000Z"
          },
          "phone": {
            "type": "string",
            "description": "Telefone do usuário",
            "example": "+5511999999999"
          },
          "position": {
            "type": "string",
            "description": "Posição do jogador",
            "example": "DEFENDER",
            "enum": [
              "DEFENDER",
              "MIDFIELDER",
              "FORWARD"
            ],
            "nullable": true
          },
          "observations": {
            "type": "string",
            "description": "Observações sobre o usuário",
            "example": "Senha temporária gerada: abc123",
            "nullable": true
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "type",
          "createdAt",
          "updatedAt",
          "phone",
          "position",
          "observations"
        ]
      },
...
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email do usuário",
            "example": "usuario@exemplo.com"
          },
          "password": {
            "type": "string",
            "description": "Senha do usuário",
            "example": "senha123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email do usuário",
            "example": "usuario@exemplo.com"
          },
          "name": {
            "type": "string",
            "description": "Nome do usuário",
            "example": "Nome do Usuário"
          },
          "password": {
            "type": "string",
            "description": "Senha do usuário. Se não informada, uma senha temporária será gerada automaticamente",
            "example": "senha123"
          },
          "type": {
            "type": "string",
            "description": "Tipo de usuário",
            "example": "PLAYER",
            "enum": [
              "PLAYER",
              "ADMIN"
            ],
            "default": "PLAYER"
          }
        },
        "required": [
          "email",
          "name",
          "type"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "description": "Nova senha do usuário",
            "example": "nova-senha123"
          }
        },
        "required": [
          "password"
        ]
      },
...
    }
  }
}